requirements.txt
numpy==1.24.3
pandas==2.0.1
requests==2.31.0
python-dotenv==1.0.0
pydantic>=2.10.3
loguru==0.7.0
python-dateutil==2.8.2
aiohttp==3.8.5
beautifulsoup4==4.12.2
lxml==4.9.3
web3==6.15.1
firecrawl==1.12.0
anthropic==0.18.1
mem0ai==0.1.60 